+++ src/losant/ensure-starts-with.test.js
@@ -1,6 +1,6 @@
 const { pipe, map, apply, adjust } = require('ramda');
 
-const testHarnessUnary = require('./test-harness-unary');
+const { testHarnessUnary } = require('../../test/utils');
 const ensureStartsWith = require('./ensure-starts-with');
 
 const cases = [