+++ src/losant/update-key-paths.test.js
@@ -1,6 +1,6 @@
 const { map, apply } = require('ramda');
 
-const testHarnessUnary = require('./test-harness-unary');
+const { testHarnessUnary } = require('../../test/utils');
 const updateKeyPaths = require('./update-key-paths');
 
 const updateFooKeyPath = updateKeyPaths({ foo: ['params', 'bar'] });